Extract out pointer input handling
We have now split out the pointer input handling to 2 sides: raw event
handling which maps raw events to higher-level events
(PagerPointerEvent), and then a receiving side which updates the UI as
required from the PagerPointerEvents.

Internally this is done through a Channel, allowing appropriate queueing
on both the sending and receiving sides.
