+++ pager/src/main/java/dev/chrisbanes/accompanist/pager/PagerState.kt
@@ -1,0 +1,327 @@
+/*
+ * Copyright 2021 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@file:Suppress("MemberVisibilityCanBePrivate")
+
+package dev.chrisbanes.accompanist.pager
+
+import android.util.Log
+import androidx.annotation.FloatRange
+import androidx.compose.animation.core.AnimationSpec
+import androidx.compose.animation.core.AnimationState
+import androidx.compose.animation.core.DecayAnimationSpec
+import androidx.compose.animation.core.animate
+import androidx.compose.animation.core.animateDecay
+import androidx.compose.animation.core.calculateTargetValue
+import androidx.compose.animation.core.spring
+import androidx.compose.foundation.MutatePriority
+import androidx.compose.foundation.MutatorMutex
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+import kotlinx.coroutines.CancellationException
+import kotlinx.coroutines.channels.ReceiveChannel
+import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.isActive
+import kotlinx.coroutines.launch
+import kotlin.math.absoluteValue
+import kotlin.math.floor
+import kotlin.math.roundToInt
+
+/**
+ * A state object that can be hoisted to control and observe scrolling for [Pager].
+ *
+ * @param currentPage the initial value for [PagerState.currentPage]
+ * @param pageCount the initial value for [PagerState.pageCount]
+ */
+class PagerState(
+    currentPage: Int = 0,
+    pageCount: Int = 0,
+) {
+    private var _pageCount by mutableStateOf(pageCount)
+
+    /**
+     * The index of the currently selected page.
+     */
+    var pageCount: Int
+        get() = _pageCount
+        set(value) {
+            _pageCount = value.coerceAtLeast(0)
+            currentPage = currentPage.coerceIn(0, pageCount)
+        }
+
+    private var _currentPage by mutableStateOf(currentPage.coerceIn(0, pageCount))
+
+    /**
+     * The index of the currently selected page.
+     *
+     * To update the scroll position, use [scrollToPage] or [animateScrollToPage].
+     */
+    var currentPage: Int
+        get() = _currentPage
+        private set(value) {
+            _currentPage = value.coerceIn(0, _pageCount)
+        }
+
+    private val _currentPageOffset = mutableStateOf(0f)
+
+    /**
+     * The current offset from the start of [currentPage], as a fraction of the page width.
+     *
+     * To update the scroll position, use [scrollToPage] or [animateScrollToPage].
+     */
+    @get:FloatRange(from = 0.0, to = 1.0)
+    var currentPageOffset: Float
+        get() = _currentPageOffset.value
+        private set(value) {
+            _currentPageOffset.value = value.coerceIn(
+                minimumValue = if (currentPage == pageCount) 0f else -1f,
+                maximumValue = if (currentPage == 0) 0f else 1f,
+            )
+        }
+
+    internal var pageSize by mutableStateOf(0)
+
+    /**
+     * Represents the current selection state of a [Pager].
+     * Usually read from [PagerState.selectionState].
+     */
+    enum class SelectionState {
+        Selected,
+        Undecided
+    }
+
+    /**
+     * The current selection state.
+     */
+    var selectionState by mutableStateOf(SelectionState.Selected)
+        private set
+
+    private val mutatorMutex = MutatorMutex()
+
+    /**
+     * Animate (smooth scroll) to the given page.
+     *
+     * Cancels the currently running scroll, if any, and suspends until the cancellation is
+     * complete.
+     *
+     * @param page the page to snap to. Must be between 0 and [pageCount] (inclusive).
+     * @param pageOffset the percentage of the page width to offset, from the start of [page]
+     * @param initialVelocity Initial velocity in pixels per second, or `0f` to not use a start velocity.
+     */
+    suspend fun animateScrollToPage(
+        page: Int,
+        @FloatRange(from = 0.0, to = 1.0) pageOffset: Float = 0f,
+        initialVelocity: Float = 0f,
+    ) {
+        if (page == currentPage) return
+
+        try {
+            mutatorMutex.mutate {
+                selectionState = SelectionState.Undecided
+                animateToPage(
+                    page = page.coerceIn(0, pageCount),
+                    pageOffset = pageOffset.coerceIn(0f, 1f),
+                    initialVelocity = initialVelocity,
+                )
+                selectionState = SelectionState.Selected
+            }
+        } catch (e: CancellationException) {
+            // If we're cancelled, snap to the target page
+            currentPage = page
+            currentPageOffset = pageOffset
+            selectionState = SelectionState.Selected
+
+            throw e
+        }
+    }
+
+    /**
+     * Instantly brings the item at [page] to the middle of the viewport, offset by [pageOffset]
+     * percentage of page width.
+     *
+     * Cancels the currently running scroll, if any, and suspends until the cancellation is
+     * complete.
+     *
+     * @param page the page to snap to. Must be between 0 and [pageCount] (inclusive).
+     * @param pageOffset the percentage of the page width to offset, from the start of [page]
+     */
+    suspend fun scrollToPage(
+        page: Int,
+        @FloatRange(from = 0.0, to = 1.0) pageOffset: Float = 0f,
+    ) {
+        mutatorMutex.mutate {
+            currentPage = page
+            currentPageOffset = pageOffset
+            selectionState = SelectionState.Selected
+        }
+    }
+
+    private fun snapToNearestPage() {
+        if (DebugLog) {
+            Log.d("Pager", "snapToNearestPage. currentPage:$currentPage, offset:$currentPageOffset")
+        }
+        currentPage -= currentPageOffset.roundToInt()
+        currentPageOffset = 0f
+        selectionState = SelectionState.Selected
+    }
+
+    private suspend fun animateToPage(
+        page: Int,
+        pageOffset: Float = 0f,
+        animationSpec: AnimationSpec<Float> = spring(),
+        initialVelocity: Float = 0f,
+    ) {
+        animate(
+            initialValue = currentPage + currentPageOffset,
+            targetValue = page + pageOffset,
+            initialVelocity = initialVelocity,
+            animationSpec = animationSpec
+        ) { value, _ ->
+            currentPage = floor(value).toInt()
+            currentPageOffset = currentPage - value
+        }
+    }
+
+    private fun scrollByInternal(x: Float) {
+        currentPageOffset += x / pageSize.coerceAtLeast(1)
+    }
+
+    internal suspend fun scrollBy(x: Float) {
+        mutatorMutex.mutate(MutatePriority.UserInput) {
+            scrollByInternal(x)
+        }
+    }
+
+    private fun determineSpringBackOffset(
+        offset: Float = currentPageOffset
+    ): Float = when {
+        // If the offset exceeds the scroll threshold (in either direction), we want to
+        // move to the next/previous item
+        offset >= ScrollThreshold -> 1f
+        offset <= -ScrollThreshold -> -1f
+        // Otherwise we snap-back to 0
+        else -> 0f
+    }.coerceIn(
+        minimumValue = if (currentPage == pageCount) 0f else -1f,
+        maximumValue = if (currentPage == 0) 0f else 1f
+    )
+
+    /**
+     * TODO make this public?
+     */
+    private suspend fun fling(
+        initialVelocity: Float,
+        animationSpec: DecayAnimationSpec<Float>,
+    ) {
+        // We calculate the target offset using pixels, rather than using the offset
+        val targetOffset = animationSpec.calculateTargetValue(
+            initialValue = currentPageOffset * pageSize,
+            initialVelocity = initialVelocity
+        ) / pageSize
+
+        if (DebugLog) {
+            Log.d(
+                "Pager",
+                "fling. velocity:$initialVelocity, " +
+                    "page: $currentPage, " +
+                    "offset:$currentPageOffset, " +
+                    "targetOffset: $targetOffset"
+            )
+        }
+
+        // If the animation can naturally end outside of current page bounds, we will
+        // animate with decay.
+        if (targetOffset.absoluteValue >= 1) {
+            // Animate with the decay animation spec using the fling velocity
+            AnimationState(
+                initialValue = currentPageOffset * pageSize,
+                initialVelocity = initialVelocity
+            ).animateDecay(animationSpec) {
+                // The property will coerce the value to the correct range
+                currentPageOffset = value / pageSize
+
+                if (currentPageOffset.absoluteValue >= 1) {
+                    // If we reach the bounds of the allowed offset, cancel the animation
+                    cancelAnimation()
+                }
+            }
+            snapToNearestPage()
+        } else {
+            // Otherwise we animate to the next item, or spring-back depending on the offset
+            animate(
+                initialValue = currentPageOffset,
+                targetValue = determineSpringBackOffset(targetOffset.coerceIn(-1f, 1f)),
+                initialVelocity = initialVelocity / pageSize.coerceAtLeast(1),
+                animationSpec = spring()
+            ) { value, _ ->
+                currentPageOffset = value
+            }
+            snapToNearestPage()
+        }
+    }
+
+    internal suspend fun receiveDragEvents(
+        channel: ReceiveChannel<PagerPointerEvent>,
+        flingSpec: DecayAnimationSpec<Float>,
+    ) = coroutineScope {
+        while (isActive) {
+            // Receive our first event
+            var event = channel.receive()
+
+            if (event !is PagerPointerEvent.Down) {
+                // If this the first event is 'Down', skip
+                continue
+            }
+
+            event = channel.receive()
+            try {
+                mutatorMutex.mutate {
+                    while (event is PagerPointerEvent.Drag) {
+                        scrollByInternal((event as PagerPointerEvent.Drag).dx)
+                        event = channel.receive()
+                    }
+                }
+            } catch (e: CancellationException) {
+                // TODO: cancellation, while dragging
+            }
+
+            if (event is PagerPointerEvent.Up) {
+                launch {
+                    try {
+                        mutatorMutex.mutate {
+                            fling(
+                                initialVelocity = (event as PagerPointerEvent.Up).velocity.x,
+                                animationSpec = flingSpec
+                            )
+                        }
+                    } catch (e: CancellationException) {
+                        // TODO: cancellation, while dragging
+                    }
+                }
+            } else if (event is PagerPointerEvent.Cancel) {
+                // TODO:
+            }
+        }
+    }
+
+    override fun toString(): String = "PagerState(" +
+        "pageCount=$pageCount, " +
+        "currentPage=$currentPage, " +
+        "selectionState=$selectionState, " +
+        "currentPageOffset=$currentPageOffset" +
+        ")"
+}
